+++ export_stockinfo_xls/doc/RELEASE_NOTES.md
@@ -2,8 +2,16 @@
 
 #### 01.11.2019
 #### Version 13.0.1.0.0
-#### Migration
-Migration Of Export Product Stock in Excel
+#### ADD
+Initial Commit  Export Product Stock in Excel
+
+
+#### 05.12.2019
+#### Version 13.0.1.1.0
+#### FIX
+Updation Of Export Product Stock in Excel
+
+